requirements.txt
gitpython
mock
nose
hypothesis
coverage
responses
